put login, logout and signup at one file
