+++ server/src/controllers/index.js
@@ -1,8 +1,8 @@
 const { clientError, serverError } = require('./error');
-const signup = require('./signup');
-const login = require('./login');
+const { signup, login, logout } = require('./auth');
+// const login = require('./login');
 const { getRBookingbyDate, bookingRoom, deleteBooking } = require('./booking');
-const logout = require('./logout');
+// const logout = require('./logout');
 const { addRoom, getRooms, patchRoom, deleteRoomById } = require('./room');
 const { deleteUser, getUsers, getProfile, patchProfile } = require('./users');
 const activateAccount = require('./activateAccount');
