--- server/src/controllers/logout.js
@@ -1,3 +1,0 @@
-module.exports = (req, res) => {
-  res.clearCookie('token').end();
-};
